Prepare for testing celery tasks without redis
