+++ meetupselector/celery.py
@@ -1,10 +1,7 @@
-import os
-
 from celery import Celery
 from django.conf import settings
 
 # set the default Django settings module for the 'celery' program.
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "meetupselector.settings.prod")
 app = Celery("meetupselector")
 
 # Using a string here means the worker will not have to
